Move placeholder utils to placeholder.ts. Make function names clearer and abstract logic into helper functions. Change up ordering and placement of functions for better flow
